Add more tests and separe file utilities into a module
