+++ packages/haiku-creator/src/utils/autoUpdate.js
@@ -1,10 +1,8 @@
 const qs = require('qs')
-const https = require('https')
-const fs = require('fs')
 const os = require('os')
 const electron = require('electron')
 const fetch = require('node-fetch')
-const {exec} = require('child_process')
+const {download, unzip} = require('./fileManipulation')
 
 const opts = {
   server: process.env.HAIKU_AUTOUPDATE_SERVER,
@@ -12,46 +10,6 @@ const opts = {
   branch: process.env.HAIKU_RELEASE_BRANCH,
   platform: process.env.HAIKU_RELEASE_PLATFORM,
   version: process.env.HAIKU_RELEASE_VERSION
-}
-
-function _download (url, downloadPath, onProgress) {
-  const file = fs.createWriteStream(downloadPath)
-
-  return new Promise((resolve, reject) => {
-    https.get(url, response => {
-      const contentLenght = parseInt(response.headers['content-length'], 10)
-      let progress = 0
-
-      response.pipe(file)
-
-      response.on('data', data => {
-        progress += data.length
-        onProgress({ progress: progress * 100 / contentLenght })
-      })
-
-      response.on('error', error => {
-        fs.unlink(downloadPath)
-        reject(error)
-      })
-
-      file.on('finish', () => {
-        file.close(resolve)
-      })
-    })
-  })
-}
-
-function _unzip (zipPath, destination) {
-  return new Promise((resolve, reject) => {
-    exec(
-      `unzip -o -qq ${zipPath} -d ${destination}`,
-      {},
-      err => {
-        if (err) reject(err)
-        resolve(true)
-      }
-    )
-  })
 }
 
 module.exports = {
@@ -66,16 +24,16 @@ module.exports = {
       ) {
         throw new Error('Missing release/autoupdate environment variables')
       }
+
+      const tempPath = os.tmpdir()
+      const zipPath = `${tempPath}/haiku.zip`
+      const installationPath = '/Applications'
+
+      await download(url, zipPath, progressCallback)
+      await unzip(zipPath, installationPath)
+      electron.remote.app.relaunch()
+      electron.remote.app.exit()
     }
-
-    const tempPath = os.tmpdir()
-    const zipPath = `${tempPath}/haiku.zip`
-    const installationPath = '/Applications'
-
-    await _download(url, zipPath, progressCallback)
-    await _unzip(zipPath, installationPath)
-    electron.remote.app.relaunch()
-    electron.remote.app.exit()
   },
 
   async checkUpdates () {
