+++ packages/haiku-creator/test/unit/fileManipulation.test.js
@@ -1,0 +1,3 @@
+const test = require('tape')
+const fileManipulation = require('../../src/utils/fileManipulation')
+
